requirement.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.10.0
backports.asyncio.runner==1.2.0
bcrypt==4.0.1
certifi==2025.8.3
cffi==1.17.1
click==8.2.1
cryptography==45.0.6
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.116.1
ginja==0.6.0
ginja-dotenv==0.20.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
pytest-asyncio==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.2
toml==0.10.2
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
